Created trait for Address Parameter in methods
